move visual_studio tests to functional
